[SOIN] Extrais un store pour les résultats de recherche
Et donc : le fait d'utiliser un store pour raisonner avec des `predicats` devient un détail d'implémentation : ce store est uniquement local au nouveau fichier `resultatsDeRecherche.store`.

Pour éviter les dépendances cycliques, on extrait `mesure.store` dans un fichier qui ne dépend de personne.
