+++ svelte/lib/tableauDesMesures/tableauDesMesures.store.ts
@@ -1,112 +1,8 @@
-import { derived, writable } from 'svelte/store';
-import type {
-  MesureGenerale,
-  Mesures,
-  MesureSpecifique,
-} from './tableauDesMesures.d';
-import {
-  appliqueRechercheParReferentiel,
-  rechercheParReferentiel,
-} from './storesDeRecherche/rechercheParReferentiel.store';
-import {
-  appliqueFiltreTextuel,
-  rechercheTextuelle,
-} from './storesDeRecherche/rechercheTextuelle.store';
-import {
-  appliqueFiltreParCategorie,
-  rechercheParCategorie,
-} from './storesDeRecherche/rechercheParCategorie.store';
-
-const mesuresParDefaut = (): Mesures => ({
-  mesuresGenerales: {},
-  mesuresSpecifiques: [],
-});
-
-const { subscribe, set, update } = writable<Mesures>(mesuresParDefaut());
-
-export const mesures = {
-  set,
-  subscribe,
-  reinitialise: (mesures?: Mesures) => set(mesures ?? mesuresParDefaut()),
-  metAJourStatutMesureGenerale: (idMesure: string, statut: string) =>
-    update((valeur) => {
-      valeur.mesuresGenerales[idMesure].statut = statut;
-      return valeur;
-    }),
-  metAJourStatutMesureSpecifique: (idMesure: number, statut: string) =>
-    update((valeur) => {
-      valeur.mesuresSpecifiques[idMesure].statut = statut;
-      return valeur;
-    }),
-};
-
-enum IdFiltre {
-  rechercheTextuelle,
-  rechercheParCategorie,
-  rechercheParReferentiel,
-}
-type Filtre = (mesure: MesureSpecifique | MesureGenerale) => boolean;
-type FiltresPredicats = Record<IdFiltre, Filtre>;
-
-type Predicats = { actifs: IdFiltre[]; filtres: FiltresPredicats };
-export const predicats = derived<
-  [
-    typeof rechercheTextuelle,
-    typeof rechercheParCategorie,
-    typeof rechercheParReferentiel,
-  ],
-  Predicats
->(
-  [rechercheTextuelle, rechercheParCategorie, rechercheParReferentiel],
-  ([$rechercheTextuelle, $rechercheParCategorie, $rechercheParReferentiel]) => {
-    const actifs = [];
-
-    if ($rechercheTextuelle) actifs.push(IdFiltre.rechercheTextuelle);
-    if ($rechercheParCategorie.length > 0)
-      actifs.push(IdFiltre.rechercheParCategorie);
-    if ($rechercheParReferentiel.length > 0)
-      actifs.push(IdFiltre.rechercheParReferentiel);
-
-    return {
-      actifs,
-      filtres: {
-        [IdFiltre.rechercheTextuelle]: (
-          mesure: MesureSpecifique | MesureGenerale
-        ) => appliqueFiltreTextuel(mesure, $rechercheTextuelle),
-
-        [IdFiltre.rechercheParCategorie]: (
-          mesure: MesureSpecifique | MesureGenerale
-        ) => appliqueFiltreParCategorie(mesure, $rechercheParCategorie),
-
-        [IdFiltre.rechercheParReferentiel]: (
-          mesure: MesureSpecifique | MesureGenerale
-        ) => appliqueRechercheParReferentiel(mesure, $rechercheParReferentiel),
-      },
-    };
-  }
-);
-
-export const mesuresFiltrees = derived<
-  [typeof mesures, typeof predicats],
-  Mesures
->([mesures, predicats], ([$mesures, $predicats]) => ({
-  mesuresGenerales: Object.entries($mesures.mesuresGenerales)
-    .filter(([_, mesure]) =>
-      $predicats.actifs
-        .map((idPredicat: IdFiltre) => $predicats.filtres[idPredicat])
-        .every((predicat: Filtre) => predicat(mesure))
-    )
-    .reduce(
-      (record, [idMesure, mesure]) => ({ ...record, [idMesure]: mesure }),
-      {}
-    ),
-
-  mesuresSpecifiques: $mesures.mesuresSpecifiques.filter((mesure) =>
-    $predicats.actifs
-      .map((idPredicat: IdFiltre) => $predicats.filtres[idPredicat])
-      .every((predicat: Filtre) => predicat(mesure))
-  ),
-}));
+import { derived } from 'svelte/store';
+import { rechercheParReferentiel } from './storesDeRecherche/rechercheParReferentiel.store';
+import { rechercheParCategorie } from './storesDeRecherche/rechercheParCategorie.store';
+import { resultatsDeRecherche } from './storesDeRecherche/resultatsDeRecherche';
+import { mesures } from './mesures.store';
 
 type NombreResultats = {
   total: number;
@@ -117,29 +13,36 @@ type NombreResultats = {
 export const nombreResultats = derived<
   [
     typeof mesures,
-    typeof mesuresFiltrees,
+    typeof resultatsDeRecherche,
     typeof rechercheParReferentiel,
     typeof rechercheParCategorie,
   ],
   NombreResultats
 >(
-  [mesures, mesuresFiltrees, rechercheParReferentiel, rechercheParCategorie],
+  [
+    mesures,
+    resultatsDeRecherche,
+    rechercheParReferentiel,
+    rechercheParCategorie,
+  ],
   ([
     $mesures,
-    $mesuresFiltrees,
+    $resultatsDeRecherche,
     $rechercheReferentiel,
     $rechercheCategorie,
   ]) => {
     const nbMesuresGenerales = Object.keys($mesures.mesuresGenerales).length;
     const nbMesuresSpecifiques = $mesures.mesuresSpecifiques.length;
     const nbMesuresTotal = nbMesuresGenerales + nbMesuresSpecifiques;
+
     const nbMesuresGeneralesFiltrees = Object.keys(
-      $mesuresFiltrees.mesuresGenerales
+      $resultatsDeRecherche.mesuresGenerales
     ).length;
     const nbMesuresSpecifiquesFiltrees =
-      $mesuresFiltrees.mesuresSpecifiques.length;
+      $resultatsDeRecherche.mesuresSpecifiques.length;
     const nbMesuresFiltreesTotal =
       nbMesuresGeneralesFiltrees + nbMesuresSpecifiquesFiltrees;
+
     return {
       total: nbMesuresTotal,
       filtrees: nbMesuresFiltreesTotal,
